Add special handling for missing getattr on StepData.
If a recipe accesses an attribute of a StepData object that doesn't exist, it
results in an ugly and unhelpful stacktrace. It makes it look like the recipe
failed, even though actually (usually) the preceeding step failed and the
recipe just didn't know it yet.

This CL doesn't change the fundamental behavior -- the recipe will still fail
with an exception, because the recipe should probably have aborted as soon as
the underlying step failed, and if it didn't it should be ready to handle corner
cases like these -- but it does make the resulting exception much more
reasonable for the casual observer to understand.

R=iannucci@chromium.org, luqui@chromium.org
BUG=288051

Review URL: https://codereview.chromium.org/1176603003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295606 0039d316-1c4b-4281-b951-d872f2087c98