Fix factory definitions in master_gen.py
In testing the change to always use the slave's version
of master.cfg, I realized that master_gen.py was not
setting the factory properties for the master properly.

As this only applied to builders.pyl-based bots, and
annotated run would pick up the properties directly from
builders.pyl instead, this didn't really cause any issues,
but with the new annotated_run code we need this to be
correct (and it is a good idea anyway).

R=luqui@chromium.org

Review URL: https://codereview.chromium.org/1177053003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295635 0039d316-1c4b-4281-b951-d872f2087c98