Enable goma for all NaCl builds.
This is the 4th revival of r294339.

The previous failure was that archive contained 0 byte object file
even after re-archiving objects.

In this time, when we found 0 byte object in the archive file,
we recompile it without goma, and archive. According to my
experiment on goma canary buildbot, this fixed the problem.

BUG=390764

Review URL: https://codereview.chromium.org/1178243003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295650 0039d316-1c4b-4281-b951-d872f2087c98