Modify annotated_run to read factory configs from the slave checkout.
Currently if you wish to change the recipe a builder uses, or
change the other factory properties associated with a builder,
you need to restart the master.

This change modifies annotated_run.py so that instead of taking
the information the master hands us, we ignore it and use the
copy of the master config we find in the slave checkout, which
should be up-to-date.

This should allow us to eliminate a decent number of master restarts.
(Though it would still be a good idea to restart the masters
periodically so that they don't get too stale).

R=iannucci@chromium.org, phajdan.jr@chromium.org, luqui@chromium.org
BUG=499071

Review URL: https://codereview.chromium.org/1178733003.

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295684 0039d316-1c4b-4281-b951-d872f2087c98