Add swarming support for MB (and the GN bots) in the chromium recipes.
This patch adds support for swarming to bots that use the chromium
recipes (chromium.py, chromium_trybot.py) and the MB build wrapper
(so that we can use GN).

It does this by adding a hook to the run_mb() call that takes a
list of the gtest_test names (which need to match their ninja
build target names) and passes that to `mb gen --swarming-targets`
via a temp file.

MB then re-maps the ninja target names to the GN labels, and
writes a file containing the labels into the build directory,
and passes that to GN (or, it will ignore this in a GYP build).

Based on Dirk Pranke's CL https://codereview.chromium.org/1177593002

BUG=480053

Review URL: https://codereview.chromium.org/1187773004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295711 0039d316-1c4b-4281-b951-d872f2087c98