Move Mac GN bots from the CQ pool to the optional pool.
We're close to the limit on resources for the existing Mac CQ
(non-experimental) bots, and so we want to make sure they are not
competing for builds with any optional bots.

Also, normally we don't put any non-CQ bots in the CQ pool
anyway. These were there originally because at least the non-upload
bots were supposed to go into the CQ immediately at the time,
but we had to back that change out.

R=jam@chromium.org
BUG=439402

Review URL: https://codereview.chromium.org/1176163003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295726 0039d316-1c4b-4281-b951-d872f2087c98