Fix the chromium_gn recipe to call mb after calling analyze.
As part of enabling MB and swarming for the GN bots in
the chromium and chromium_trybot recipes, we changed
things so that we call `mb analyze` before we call `mb gen`
(in that recipe). Unfortunately, in order to do that safely,
`mb analyze` needs to call `mb gen`, but doesn't get all
of the arguments it needs (in particular, it doesn't get
the --goma-dir argument). So, if `mb analyze` is called
*after `mb gen` (like the chromium_gn recipe was doing),
it would corrupt the build dir.

I need to fix analyze to make sure it gets the same arguments
gen gets, but in the meantime, we also need to make sure
chromium_gn runs the two commands in the same order
chromium_trybot does.

TBR=nodir@chromium.org,phajdan.jr@chromium.org
BUG=502091, 481693

Review URL: https://codereview.chromium.org/1194743004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295761 0039d316-1c4b-4281-b951-d872f2087c98