Enabling step nesting for reference values and revision range.
This change uses the newly added step nesting capability to make the output of
the bisect recipe more readable.

This is the first of two or three changes that enable such usage. We are
splitting it up so that it is easier to review and test.

BUG=477045
R=prasadv

Review URL: https://codereview.chromium.org/1188703003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@295790 0039d316-1c4b-4281-b951-d872f2087c98