Set up one column per template type on the tryserver waterfall.
Instead of having columns for every builder, I've now set it
up so that we just have one column for all 'paladin' builders, one for all 'pre-cq' builders, etc. This dramatically reduces the width of the waterfall and improves performance of the waterfall
in general.

Currently the waterfall can't start due to too many columns
so this is being put in as a temporary measure by the build
deputy. It can be reverted if needed.

BUG=chromium:497958
TEST=Setup local tryserver to test, launch tryjobs.
TBR=dnj

Review URL: https://codereview.chromium.org/1166183003 .

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296020 0039d316-1c4b-4281-b951-d872f2087c98