Removed all traces of include_deps. No more recursive config setting.
This CL removes the auto recursive functionality of the config module.
Previously, whenever you called set_config on a module, it automatically
would call set_config recursively on *every* dependency. This is some
pretty deep magic, which in retrospect was a bad design decision. This CL
removes this behavior, and forces you to explicitly list which dependencies
you want to configure.

This CL is part of a process to kill the current config system and provide
a new module (not integrated tightly into the recipe engine) which provides
the same functionality but in a much saner way.

The private version of this CL is here: https://chromereviews.googleplex.com/228917013


BUG=285300

Review URL: https://codereview.chromium.org/1244613003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296056 0039d316-1c4b-4281-b951-d872f2087c98