Add nacl recipe and a builder to test it with
This is the first step in (eventually) converting nacl to use recipes.
This CL adds a nacl recipe which runs the build script in the nacl
repo that builds nacl on builders (buildbot_selector.py).

(Eventually, after recipe code can live in a project's repo instead of
the infra repo, this *should* be replaced with a recipe living in the
nacl repo.)

BUG=

Review URL: https://codereview.chromium.org/1256083004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296117 0039d316-1c4b-4281-b951-d872f2087c98