Switch *all* NaCl Windows builers to recipes
The parameters passed to the remaining NaCl Windows builders were
gclient-specific. The recipe uses bot_update, and so these are not
expected to be relevant any more.

This CL switches the remaining NaCl Windows builders (a
non-parametrised batch has already been switched in a prior CL)
under the presumption that the parameters do not matter due to the
switch to bot_update, and that bot_update shall automatically ensure a
correct checkout.

BUG=
R=bradnelson@chromium.org

Review URL: https://codereview.chromium.org/1256293004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296136 0039d316-1c4b-4281-b951-d872f2087c98