Re-land "Set up one column per template type on the tryserver waterfall."
Instead of having columns for every builder, I've now set it
up so that we just have one column for all 'paladin' builders,
one for all 'pre-cq' builders, etc. This dramatically reduces
the width of the waterfall and improves performance of the waterfall
in general.

BUG=chromium:497958
TEST=Setup local tryserver to test, launch tryjobs, verify cbuildbot step is
     there.

R=dnj@chromium.org

Review URL: https://codereview.chromium.org/1260043003 .

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296161 0039d316-1c4b-4281-b951-d872f2087c98