Adds a whitelist for __init__ to the RecipeApiMeta wrapping code.
This CL adds a whitelist for the __init__ method for all RecipeApiMeta
subclasses. This solves a problem I was running into where I couldn't
see the arguments for the __init__ method through reflection. This causes
the __init__ method to not be whitelisted, so that I can see argument names.

R=iannucci@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1265833003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296177 0039d316-1c4b-4281-b951-d872f2087c98