Changed incremental report to use the json module.
Changed incremental_coverage_report to output results using the JSON
recipe module instead of uploading to a storage bucket with gsutil. This saves an error
handling step in the case that the output file is empty and makes more
sense since old reports do not need to be archived.

BUG=501536

Review URL: https://codereview.chromium.org/1271913002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296208 0039d316-1c4b-4281-b951-d872f2087c98