Make annotated_run's 'builder not found' accurate
Before this change, annotated_run.py presumed that if a builder had no
properties, that builder didn't exist. This changes that check so that
it does not implicitly presume that an empty properties list implies an
absent builder, but instead checks explicity for whether a builder is
found or not.

BUG=None
R=bradnelson@chromium.org

Review URL: https://codereview.chromium.org/1281513004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296249 0039d316-1c4b-4281-b951-d872f2087c98