Change MB step to use the real buildername for tryservers.
When running MB as part of the chromium_trybot recipe, we
were previously using the mastername and buildername of
the waterfall bot the tryserver was configured to match.

This didn't work properly, because in some cases we actually
use different GYP_DEFINES on the tryserver and waterfall bot
(for example, to enable DCHECKs in release), and we need to
have separate entries in MB for each config.

TBR=phajdan.jr@chromium.org
BUG=481692

Review URL: https://codereview.chromium.org/1278793008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296260 0039d316-1c4b-4281-b951-d872f2087c98