Always resolve commit position from origin instead of local master.
In bot_update when a patch is applied the branch is detached (HEAD detached at origin/master).
The current logic for resolving commit position looks for git hashes in ref/heads/master,
if the master is not up-to-date, then this method will not be able to resolve commit-position.
BUG=517631

Review URL: https://codereview.chromium.org/1278023002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296268 0039d316-1c4b-4281-b951-d872f2087c98