re-land r296260 and use real builder names when running MB.
The previous version of this change was okay; it was reverted along
with the change to use MB on more bots for consistency.

No expectations change in this CL because none of the bots currently
using MB had different results based on the builder they were matching.

TBR=phajdan.jr@chromium.org
BUG=481692

Review URL: https://codereview.chromium.org/1284253002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296276 0039d316-1c4b-4281-b951-d872f2087c98