Re-land r296259 and try again to use MB on the Chromium Linux builders.
This should affect the following bots:

on chromium.linux:
- Linux Builder
- Linux Tests
- Linux Builder (dbg)
- Linux Tests (dbg)(1)

on tryserver.chromium.linux:
- linux_chromium_compile_dbg_ng
- linux_chromium_dbg_ng
- linux_chromium_rel_ng

on client.skia:
- Linux Builder
- Linux Builder (Trybot)
- Linux Tests
- Linux Tests (Trybot)

The tryserver and skia bots are affected because they attempt to
match the waterfall bots' configs.

Re-landing this patch w/o changes as there was (we think) nothing
wrong with this patch the first time, the bugs were are repo-side.

TBR=phajdan.jr@chromium.org
BUG=481692

Review URL: https://codereview.chromium.org/1288783002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296284 0039d316-1c4b-4281-b951-d872f2087c98