Add config for an opt-in CFI try bot.
CFI stands for Control Flow Integrity, see
https://sites.google.com/a/chromium.org/dev/developers/testing/control-flow-integrity

There's already a CFI buildbot:
build.chromium.org/p/chromium.fyi/builders/CFI Linux

This CL adds a try bot to minimize breakages of the buildbot.
The try bot is opt-in, because it's only intended to be used,
when new test suites are about to be added to the CFI buildbot.
Otherwise, the new CFI failures which be caught by the buildbot
(they are very rare; although, often very serious).

BUG=chromium:457523

Review URL: https://codereview.chromium.org/1281293003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296290 0039d316-1c4b-4281-b951-d872f2087c98