Always set GYP_DEFINES during runhooks, even when not running gyp
There are some runhooks that peek in to GYP_DEFINES so
these need to be set during runhooks even if we are not
using gyp. For example, the cross compile bot such as
32-bit linux and arm linux will only install the sysroot
if they see target_arch in GYP_DEFINES.

This problem went unnoticed until a recent change to start
using mb.py: https://codereview.chromium.org/1292273003

In the long run we probably need to find a cleaner/better
way to do this.

BUG=521162

Review URL: https://codereview.chromium.org/1293953002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296330 0039d316-1c4b-4281-b951-d872f2087c98