Add a way to expose a list of trybots from buildbot via JSON
Some buildbot instances (e.g., GYP, Skia) support running both continuous builders and tryservers on the same instance; in that situation, there's currently no good way to expose a list of buildernames via the JSON API so that services like Rietveld can figure out which builders on the master should be listed as tryservers.

This change adds a mechanism for this. It adds a new per-builder property, 'trybot' (a bool), and modifies the JSON API entrypoint, '/json/builders' to take an optional 'trybots=1' param, that returns the same thing as '/json/builders' except the list is filtered to just the builders that set that per-builder parameter.

In order to take advantage of this entrypoint the master.cfg must be updated to set this key, and any callers (like Rietveld) must know to call this special entry point to get the list of builders.

This CL also updates the builders.pyl format to support a new corresponding, optional, per-builder key ('trybot': bool).

R=nodir
BUG=504072

Review URL: https://codereview.chromium.org/1287063009

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296433 0039d316-1c4b-4281-b951-d872f2087c98