Add new recipe for bot to run AndroidWebViewTest on appurify.
The amp device pool that the amp-split bot is connnected to only
has Nexus 5's. This is because it is a trial bot for a bot on the
CQ which will only run tests on N5s for stability reasons. Going to
make a new bot to run AndroidWebViewTest.apk on amp.


BUG=

Review URL: https://codereview.chromium.org/1309993002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296471 0039d316-1c4b-4281-b951-d872f2087c98