Making bisector accept the name of the tester to use.
Bisector will now look at the recipe_tester_name buildbot property and
use it to perform the tests. It also uses this bot name to decide which
bot to use for building and which gs path to use to look for the build
image.

This change also allows for the previous 'orginal_bot_name' logic to
preserver backwards compatibility.

R=prasadv

Review URL: https://codereview.chromium.org/1313073006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296474 0039d316-1c4b-4281-b951-d872f2087c98