Remove Linux ARM Compile from main waterfall.
The motivation for this was to ensure ARM compiles don't break and to run some NaCl browser tests on ARM. The former is handled by the chromeos_daisy bot on the CQ and main waterfall, which builds our ARM shipping configuration. For the latter (tests), we don't run tests on the CQ anyways, and those can be run through ChromeoS images instead. This bot remains on chromium.fyi.

BUG=384104

Review URL: https://codereview.chromium.org/1301393004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296475 0039d316-1c4b-4281-b951-d872f2087c98