Explicitly build 'all' on all deterministic builders.
The pseudo 'chromium_swarm_tests' should be deleted eventually and for
deterministic build, we want all the build to be deterministic. This will lead
to a regression of the deterministic builds, which will be fixed in a follow up
once we have more data.

It is done on the staging server so we can assert the current level before
enabling it to be run by default on the Try Server / CI infrastructure.

TBR=sebmarchand@chromium.org
BUG=314403

Review URL: https://codereview.chromium.org/1309543010

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296485 0039d316-1c4b-4281-b951-d872f2087c98