Syzygy recipe conversion to use properties as arguments.
This CL is part of a refactor of recipes which use the properties supplied
by buildbot. This refactor is to both expose the global state of properties
to users of the recipe system, and to start to treat recipes as "first class"
objects which have arguments and return values, in preparation for the future
move to LUCI. The properties given by buildbot are converted by the recipe
system to Python arguments, which are then passed into the recipe or recipe
module. This CL removes any references to api.properties, and replaces those
with the arguments given by the recipe system. Recipe documentation can be
found here: 
https://chromium.googlesource.com/chromium/tools/build/+/HEAD/third_party/recipe_engine/doc/user_guide.md#Properties-are-the-primary-input-for-your-recipes
CC=iannucci@chromium.org

BUG=

Review URL: https://codereview.chromium.org/1309743003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296531 0039d316-1c4b-4281-b951-d872f2087c98