Cause test_env.py to try to import the compiled coverage first.
This CL causes test_env.py to try to use the compiled version of coverage
first. Previously, it would import coverage blindly, and almost always get
the python version, then realize that, delete it, and add the compiled
version to the path. This makes it so it tries to import the compiled version
by default.

BUG=

Review URL: https://codereview.chromium.org/1323073008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296534 0039d316-1c4b-4281-b951-d872f2087c98