Converts x86 android builder on chromium.fyi to recipes.
This CL converts the android x86 builder which is currently on chromium.fyi
to use recipes. This should be a very simple change, since it's doing essentially
the same thing as the x64 builder; the only difference in their expectations
is the architecture they're compiling on.

CL for new recipe: https://codereview.chromium.org/1308393005

BUG=415873

Review URL: https://codereview.chromium.org/1308193003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296552 0039d316-1c4b-4281-b951-d872f2087c98