Choose preferred build slave if any. (Win trybot case)
Some trybots are sharing many build slaves for the different
configurations.

Accoding to shinyak's investigation, when we can reuse the same
configuration slave, we can reduce the number of necessary compiles.
So compile time can be improved, and goma executor load can be reduced.
It is estimated that we can save 20~30% of compile requests, and 10% of
goma executor load.

buildbot can configure how to choose the slave by setting 'nextSlave'
function. In this patch, we achieve this by:
1) setting 'preferred_builder' value in slave.
2) if a builder name matches the preferred_builder, the builder uses
   that builder preferrably.

In this patch, we set preferred_builder for win exp bots. exp bots will
use the same configuration slave if possible. If no preferred builder,
master will choose other slave at random.

When this succeeds, we try to set nextSlave for all CQ builders.

BUG=524581

Review URL: https://codereview.chromium.org/1315143005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296557 0039d316-1c4b-4281-b951-d872f2087c98