Switch NaCl tryserver to recipes
This is a major overhaul of the NaCl tryserver, to convert it to
recipes. Some of the changes, in no particular order:

* slaves.cfg now contains builder categories, a mapping of categories to
slaves, and another of categories to builders. These three dictinaries
are used to construct the actual list of slaves.

* master.cfg now no longer contains the mammoth list of projects and
platforms that it did; this data is now taken directly from slaves.cfg.

* The list of builders, slaves, and the mapping between them now resides
solely in one place (slaves.cfg); it isn't spread across master.cfg and
slaves.cfg as it hitherto was.

* The list of builders in slaves.cfg now consists of actual builders,
and not platform or category names.

* The convoluted logic used to construct a builder is now considerably
simpler, requiring only two factories (the recipe factories for NaCl and
NaCl ports).

* Whatever other minor changes were necessary to convert the tryserver.

BUG=
R=bradnelson@chromium.org

Review URL: https://codereview.chromium.org/1326703003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296572 0039d316-1c4b-4281-b951-d872f2087c98