Add logic to daemonizer to be able to kill what it daemonizes
Previously, daemonzer.py simply launched a daemon process without
storing its pid anywhere. Consequently, it was left to the daemon
itself to handle shutting down and/or restarting. This cl adds the
option to store the daemon's pid in a file and kill it.

This update is part of the effort to add temperature monitoring to
android devices. A more advanced daemon manager would certainly help.

BUG=519884

Review URL: https://codereview.chromium.org/1328623004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296613 0039d316-1c4b-4281-b951-d872f2087c98