Fix the --target argument passed to sizes.py.
Previously it would pass 'Release' instead of 'Release_x64' for 64-bit
Windows builds, causing sizes.py to look for files in the wrong directory.

Also implement the 'name' property for the sizes step so the recipe doesn't
crash trying to print the name if it fails.

BUG=451078

Review URL: https://codereview.chromium.org/1335793002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296641 0039d316-1c4b-4281-b951-d872f2087c98