Enable device temp monitoring on perf bots
This adds steps to the chromium_android recipe module to startup
and shutdown a device temperature monitoring daemon. Additionally, 
it enables this monitor for the android perf bots.

This change was applied to the following fyi bot overnight:
http://build.chromium.org/p/chromium.fyi/builders/Android%20Tests%20%28trial%29%28dbg%29

It generated this graph:
http://shortn/_AVd4jfRnFs

If this cl lands, that graph would start to be populated by devices on the perf waterfall as well.

BUG=519884

Review URL: https://codereview.chromium.org/1337503004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296652 0039d316-1c4b-4281-b951-d872f2087c98