Convert buildbucket property to JSON
Properties are serialized into database using __repr__, which returns Python
literal serialized into a string. In order to parse this value in Buildbot we
are forced to use unsafe ast.literal_eval. Instead, in this CL, we serialize
this value to JSON and parse it with json.loads.

R=nodir@chromium.org
CC=pgervais@chromium.org
BUG=525789

Review URL: https://codereview.chromium.org/1328623003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296698 0039d316-1c4b-4281-b951-d872f2087c98