V8 Buildbot: Fix fetching changes for auto-bisect.
Don't wrap call to pycurl.py with runit.py. This was a relic
of manual testing. In recipes, the python path is already
set. Wrapping it fails on windows, because it lacks another
call to the python executable (the python script itself
is not executable on windows).

BUG=chromium:511215
TBR=tandrii@chromium.org

Review URL: https://codereview.chromium.org/1351473002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296706 0039d316-1c4b-4281-b951-d872f2087c98