Remove 'More NaCl' builders from chromium.fyi
chromium.fyi has a number of builders that are of no significance, as
evidenced by their being in a failing state for 200 or more builds.

This CL is the first of a series that identifies and removes such
builders.

Specifically: this CL removes all the builders in the 'More NaCl'
category.

BUG=
R=iannucci@chromium.org,dnj@chromium.org

Review URL: https://codereview.chromium.org/1354053002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296751 0039d316-1c4b-4281-b951-d872f2087c98