Remove ninja bots from chromium.fyi
chromium.fyi has a number of builders that are of no significance, as
evidenced by their being in a failing state for 200 or more builds.

This CL is the third of a series that identifies and removes such
builders.

Specifically: this CL removes the 'Cr Win VS2013' and
'Cr Win VS2013 (dbg)' builders, which constitute the entirety of the
'ninja' category.

BUG=
R=dnj@chromium.org,iannucci@chromium.org

Review URL: https://codereview.chromium.org/1351193003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296768 0039d316-1c4b-4281-b951-d872f2087c98