Remove Mac OpenSSL builder in chroimum.fyi
chromium.fyi has a number of builders that are of no significance, as
evidenced by their being in a failing state for 200 or more builds.

This CL is the second of a series that identifies and removes such
builders.

Specifically: this CL removes the sole builder (Mac OpenSSL) in the
'openssl' category, which has no recent builds at all.

BUG=
R=dnj@chromium.org,iannucci@chromium.org

Review URL: https://codereview.chromium.org/1348663005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296797 0039d316-1c4b-4281-b951-d872f2087c98