V8 Buildbot: Auto-bisect - fix checking available builds.
This fixes two problems:
1) The build range on testers must always include the latest
revision as it is an assumption of bisection that the latest
revision is known to be "bad" and therefore won't be tested
again. If it's not included, another revision is wrongly
taken as the latest and the wrong culprit might be blamed.
2) 'gsutil ls' called with more than one object doesn't check
subsequent objects after the first error. That means when
checking for builds 1, 2, 3, with 2 missing, it'll only
report 1 and not 1, 3. Now each object is checked with a
separate call.

BUG=chromium:511215

Review URL: https://codereview.chromium.org/1365503004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296859 0039d316-1c4b-4281-b951-d872f2087c98