Fixing get_revisions_to_eval to get the revision in the center.
The current approach is defective, because for odd numbers of revisions
we are getting the last one instead of the one in the middle.

Note that we aim to have this function work for getting n evenly
distributed revisions along the range. Getting the revision in the
center of the range is only the base case.

R=prasadv
BUG=535682

Review URL: https://codereview.chromium.org/1366093002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296866 0039d316-1c4b-4281-b951-d872f2087c98