Add MB support to the base (gclient) buildbot factory code.
We have a few non-recipe-based bots that need to be
flipped to use GN instead of GYP now that the Linux x64
build of Chromium is using GN. This means that we
need to add support to the gclient_factory code to
know whether to call MB explicitly (and turn GYP off
in gclient runhooks) or let GYP run as part of gclient
runhooks as normal.

This will be configured by a per-builder factory property,
use_mb=True.

R=phajdan.jr@chromium.org, kerz@chromium.org
BUG=481692, 530733

Review URL: https://codereview.chromium.org/1370563003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296887 0039d316-1c4b-4281-b951-d872f2087c98