Re-land: Convert buildbucket property to JSON
Original CL: https://codereview.chromium.org/1328623003/.

Properties are serialized into database using __repr__, which returns Python                                             
literal serialized into a string. In order to parse this value in Buildbot we                                            
are forced to use ast.literal_eval. Instead, in this CL, we serialize this                                               
value to JSON and parse it with json.loads. This is more consistent with the                                             
rest of the infrastructure and makes it possible to parse it in programs
written in other languages than Python.

R=nodir@chromium.org
CC=pgervais@chromium.org
BUG=525789
NOPRESUBMIT=true

Review URL: https://codereview.chromium.org/1347143002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296904 0039d316-1c4b-4281-b951-d872f2087c98