Empty per_page_values shouldn't imply test success.
The entire test harness failed which caused per_page_values to come back empty, so that the recipes found no concrete failures, and thus assumed that there was success. Now we pay attention to the return code explicitly, so that we will report failure in that case.

Also changed default step_test_data for LocalTelemetryGPUTest to be passing instead of failing, which was nonsense.

Depends on https://codereview.chromium.org/1366353002/ (starting on patchset 13, so inter-patch diffs won't work across that boundary)

BUG=536192,499940

Review URL: https://codereview.chromium.org/1372723002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296913 0039d316-1c4b-4281-b951-d872f2087c98