[Android] Name appurify result files their test_id instead of suite name.
Currently, if you run the same test twice in a single test run (even with
different devices or other configs) the test results will overwrite each other.
This is because the results are simply saved in <suite_name>.json. Changing
the results to be saved in <test_run_id>.json for each test which will be
unique.

BUG=

Review URL: https://codereview.chromium.org/1373873002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296927 0039d316-1c4b-4281-b951-d872f2087c98