Fix bug in how we invoke MB in the gclient factory code.
I introduced a bug in r296887, where I turned off GYP in
gclient runhooks, but forgot to clear the setting from the
environment, so it stayed turned off when I later invoked
MB to generate the build files. Due to a bug in MB, that
env var propagated back to GYP, which did nothing. There
are thus two bugs to fix: one in the factory code, to ensure
that we don't pass potentially confusing flags to MB, and one
in the MB code, to ensure that we ignore the flags passed to it.

R=amineer@chromium.org
BUG=530733

Review URL: https://codereview.chromium.org/1373053003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@296934 0039d316-1c4b-4281-b951-d872f2087c98