WebRTC: Pass revision to more tests
Previously revision was only passed to perf tests, but more
tests will need this when running as isolated tests.

The chromium_android recipe module functions does not yet
accept revision as an argument so Android tests are not changed.

Passing a dummy value to --webkit-revision should work around
a hard-coded assumption of a Chromium checkout in runtest.py,
which breaks when trying to run isolated tests with Swarming.

BUG=497757

Review URL: https://codereview.chromium.org/1390473002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297005 0039d316-1c4b-4281-b951-d872f2087c98