When applying a gerrit patch, do a checkout before soft resetting.
The previous method would do a soft reset to the patch which would
generate a correct list of differing files but not update the working
tree.  Instead, checkout the patch so that it is fully applied and then soft
reset back to the base revision.

BUG=

Review URL: https://codereview.chromium.org/1387213002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297034 0039d316-1c4b-4281-b951-d872f2087c98