Run the long test suite on the Dr. Memory bots
Dr. Memory now has a short pre-commit suite for developers that is split
from a long suite, but the long suite is not very long and so rather than
double the number of slaves we simply always run the long suite on the
bots.

TBR=zhaoqin@chromium.org
BUG=https://github.com/DynamoRIO/drmemory/issues/1790
NOTRY=true

Review URL: https://codereview.chromium.org/1402443003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297118 0039d316-1c4b-4281-b951-d872f2087c98