WebRTC: Make webrtc-samples changes not trigger builds.
It turns out it is rare that webrtc-samples triggers our builds since
they're quite rare, so it's been unknown for a long time that they don't
work (the revision gets passed to bot_update, which doesn't find a commit
with that hash). 
Example: https://build.chromium.org/p/chromium.webrtc/builders/Win%20Builder/builds/29275/

This CL turns off the triggering of builds for that repo, but keeps it being
polled since it's useful for it to show up in the blamelist when figuring out
why some of our tests stop working (it has happened before).

Review URL: https://codereview.chromium.org/1394633004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297147 0039d316-1c4b-4281-b951-d872f2087c98