Don't sum "page_load_time" metrics.
I believe that it's unnecessary to sum page_load_time metrics; for some
time in the past, the summary metric was obtained by summing, but now
it is not. For example, see:
https://chromeperf.appspot.com/group_report?keys=agxzfmNocm9tZXBlcmZyFAsSB0Fub21hbHkYgIDAhcOytAkM

Even if page load time summary metrics are in some cases still obtained by summing individual
per-page results, it would be OK for bisect to treat them the same as any other metrics.

This change slightly reduces the complexity in this module by removing a special case.

Review URL: https://codereview.chromium.org/1410333008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297302 0039d316-1c4b-4281-b951-d872f2087c98