[Android] Add test timeout arg for AMP tests. Use it for WebView tests.
Lots of the test runs on AMP are timing out as they take a little
but longer than 30 minutes (the default timeout). Adding support
to specify a custom timeout and setting it to 60 minutes for the
WebView AMP tests.

Also, changing the WebView AMP configs slightly. Am now specifying
the exact device names to run on, instead of just the oems, to get
more consistency in the test runs.

BUG=

Review URL: https://codereview.chromium.org/1422073005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297374 0039d316-1c4b-4281-b951-d872f2087c98