WebRTC: Remove sizes step from chromium.webrtc.fyi bots.
The sizes step produced numbers that doesn't match the ones
for the builder in master.chromium and we haven't got any
useful data out of it for a long time. Let's just remove it and
spend time on implementing crbug.com/432052 for standalone
WebRTC if we prioritize catching static intializers early on
(which was the only reason for adding this step to these bots).

BUG=549142
TBR=phoglund@chromium.org

Review URL: https://codereview.chromium.org/1406343007

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297398 0039d316-1c4b-4281-b951-d872f2087c98