V8 Buildbot: Experiment - Slim swarming testers.
A slim swarming tester is a bot with type "tester" that runs
only swarming tests. It'll not checkout code or run hooks.
Instead it will only checkout the swarming client and
trigger/collect the tests.

The isolate hashes are provided by the triggering builder.
In the bisect case, the isolated hashes are downloaded
from the isolated json archives.

BUG=chromium:535160

Review URL: https://codereview.chromium.org/1415693009

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297422 0039d316-1c4b-4281-b951-d872f2087c98