Fix bootstrapping of subdir buildbots.
This changes the layout of subdir buildbots from /c/subdir
to build/nested/subdir. The nested folder is created on
demand for subdir slaves. Othersiwe, manual bootstrapping
is required to setup a hardcoded /c dir.

This also enforces git deps for subdir checkouts as right
now this still uses svn and fails on new slaves.

Furthermore, use an environment with depot_tools in the
PATH when syncing. Otherwise hooks fail on startup.

Also copy .boto file into the subdir checkouts. Right now
it is missing.

Also gracefully stop after spawning subdir buildbots. They
are never used together with a parent buildbot slave. If
we try to start that one as well, it fails because of a
missing active master.

BUG=chromium:535160

Review URL: https://codereview.chromium.org/1426633004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297488 0039d316-1c4b-4281-b951-d872f2087c98