Add an optional android_chromium_gn_rel trybot and rename "Blink Trybot"
This adds a bot called 'android_chromium_gn_rel' to the
tryserver.chromium.android waterfall. It can be used
for tryjobs that want to run real tests on devices. It
is not planned to be part of the CQ.

The bot will mirror the config of the 'Android GN' bot
on the chromium.linux waterfall.

This CL also renames "Blink Trybot" to "android_blink_rel" for
consistency.

Both bots will share the same slave pool.

R=jbudorick@chromium.org, stip@chromium.org, phajdan.jr@chromium.org
BUG=477777

Review URL: https://codereview.chromium.org/1420883005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297499 0039d316-1c4b-4281-b951-d872f2087c98