add support for android gce instances
This is the first cl of many to come that will add support 
for instances of android in the cloud. Using the google api client 
library,  this file will be used to launch/shutdown the android
instances before/after running tests in a build.

Left some methods unimplemented for the sake of brevity. They'll
come in a later CL.

BUG=543313

Review URL: https://codereview.chromium.org/1395923006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297534 0039d316-1c4b-4281-b951-d872f2087c98