Update recipes to use the new interface to analyze.
The new interface to analyze in GYP/MB/GN takes a list of
test_targets and additional_compile_targets as input, and
returns lists of test_targets and compile_targets as output.

Using the new interface allows us to produce more accurate results
and simplify the code in the filter and chromium_tests modules
and the calling clients.

R=phajdan.jr@chromium.org, sky@chromium.org
BUG=552146

Review URL: https://codereview.chromium.org/1438253003 .

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297571 0039d316-1c4b-4281-b951-d872f2087c98