Do not delete MB files when clobbering a build directory.
When we run compile.py --clobber, we do not want to delete
either the 'mb_type' file or the 'runtime_deps' file. If
we delete the former, the next time we do a build MB will
think it's doing a build for the first time and trigger
a second clobber. If we delete the latter, then the ninja
command to regenerate the build files won't exist, and
if we need to regenerate the build files (which shouldn't
normally be the case, but might happen if the build is
faulty) will fail.

TBR=thakis@chromium.org, phajdan.jr@chromium.org
BUG=536292, 557553

Review URL: https://codereview.chromium.org/1452343003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297602 0039d316-1c4b-4281-b951-d872f2087c98