buildbucket: fix race in clean_completed_build_requests
The purpose of clean_completed_build_requests is to find cancelled
build requests in buildbot and cancel them on buildbucket.

It used lease['build'] to determine if a build request has builds, but
in case of a lag of buildbot callbacks, in particular on_build_started,
build['lease'] may be None whlie a build has already started.

Read the build from the database instead.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1461393002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297656 0039d316-1c4b-4281-b951-d872f2087c98